Fixes bug when result set has NULL
If a query result set has a NULL for a column i,  attempting to getObject(i) gives a  null object references and this leads to a NullPointerException. The fix is to have this column value be an empty String, as that matches how the CSV file is parsed when there's no value for that column.